+++ apps/sendmail/urls.py
@@ -1,0 +1,11 @@
+
+# -*- coding: utf-8 -*-
+
+from django.conf.urls import url
+
+from .views import sendmail_send
+
+
+urlpatterns = [
+    url(r'^$', sendmail_send, name='sendmail_send'),
+]