+++ studlan/urls.py
@@ -27,6 +27,7 @@ urlpatterns = [
     url(r'^profile/', include('apps.userprofile.urls')),
     url(r'^team/', include('apps.team.urls')),
     url(r'^seating/', include('apps.seating.urls')),
+    url(r'^sendmail/', include('apps.sendmail.urls')),
     url(r'^messages/', include('postman.urls', namespace='postman')),
 
     # Views