+++ CHANGELOG.md
@@ -7,8 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Added feature for sending emails to multiple users.
 
 ### Changed
+- Replace the old email backend "django-mailgun" with "django-anymail". (Configuration compatibility code has been added to avoid breaking existing Mailgun configurations.)
 
 ### Deprecated